Fix SonarCloud Issues (#668) - zFernand0 did these, but had to amend for DCO...
* Fix sc bug


* remove duplicated lines


* forgot magic numbers


* Fix code smells


* resolve mode code smells


* Remove duplicated code


* fix lint errors


* remove more duplicated code


* remove dsActions.renameDataset (since it was already implemented in the datasetProvider)


* similar to previous commit
remove dsActions.renameDatasetMember (since it was already implemented in the datasetProvider)


* resolve more code-smells


* move cleanDir functions out of extension.ts


* refactor addZoweSession


* Move addZoweSession functions to utils and fix tests


* move moveTem[Folder function to utils


* move getSecurityModules function out of extension


* fix two more code smells
