+++ src/job/ZoweJobNode.ts
@@ -199,7 +199,7 @@ export class Job extends ZoweTreeNode implements IZoweJobTreeNode {
             try {
                 jobsInternal = await ZoweExplorerApiRegister.getJesApi(this.getProfile()).getJobsByOwnerAndPrefix(owner, prefix);
             } catch (error) {
-                await errorHandling(error, this.label, localize("getChildren.error.response", "Retrieving response from ") + `zowe.GetJobs`);
+                errorHandling(error, this.label, localize("getChildren.error.response", "Retrieving response from ") + `zowe.GetJobs`);
             }
         }
         return jobsInternal;
